Extract serializeTransaction, transactionId into serialize.ts
