+++ packages/iov-lisk/types/serialization.d.ts
@@ -1,0 +1,3 @@
+import { FullSignature, TransactionIdBytes, UnsignedTransaction } from "@iov/bcp-types";
+export declare function serializeTransaction(unsigned: UnsignedTransaction): Uint8Array;
+export declare function transactionId(unsigned: UnsignedTransaction, primarySignature: FullSignature): TransactionIdBytes;
